+++ R/parse.R
@@ -1,0 +1,57 @@
+#' Parse components of a subtitle file
+#'
+#' @param x A character vector with the lines of an `.srt` file.
+#' @param collapse The character with which to separate subtitle lines.
+#' @export
+srt_seconds <- function(x) {
+  nl <- newline(x)
+  times <- strsplit(x[c(2, nl + 2)], "\\s-->\\s")
+  times <- lapply(times, as_seconds)
+  seconds <- do.call(Map, c(f = c, times))
+  names(seconds) <- c("start", "end")
+  return(seconds)
+}
+
+#' @rdname srt_seconds
+#' @export
+srt_index <- function(x) {
+  nl <- which(x == "")
+  x <- x[c(1, nl + 1)]
+  as.integer(x[!is.na(x)])
+}
+
+#' @rdname srt_seconds
+#' @export
+srt_text <- function(x, collapse = "\n") {
+  nl <- newline(x)
+  a <- c(2, nl + 2) + 1
+  b <- c(which(x == "") - 1, length(a))
+  if (x[length(x)] == "") {
+    b[length(b)] <- b[length(b)] - 1
+  }
+  y <- rep(NA_character_, length(a))
+  for (i in seq_along(a)) {
+    y[i] <- paste(x[seq(a[i], b[i])], collapse = collapse)
+  }
+  return(y)
+}
+
+newline <- function(x) {
+  nl <- which(x == "")
+  if (x[length(x)] == "") {
+    nl <- nl[-length(nl)]
+  }
+  return(nl)
+}
+
+as_seconds <- function(y) {
+  y <- gsub(",", ".", y)
+  y <- strsplit(y, ":")
+  y <- lapply(y, FUN = as.numeric)
+  vapply(y, sum_time, double(1))
+}
+
+sum_time <- function(l) {
+  l[2] <- l[2] + (l[1] * 60)
+  l[3] + (l[2] * 60)
+}