+++ R/utils.R
@@ -1,0 +1,11 @@
+as_tibble <- function(x) {
+  if (is_installed("tibble")) {
+    tibble::as_tibble(x)
+  } else {
+    x
+  }
+}
+
+is_installed <- function(pkg) {
+  isTRUE(requireNamespace(pkg, quietly = TRUE))
+}